feat: graphQL subscriptions support (#2120)
* feat: add extra params to subscription

* feat: create hasura live provider

* feat: add liveProvider to hasura example

* refactor: fix params, types and names

* feat: add extra params to useSubscription

* feat: add error checkpoints

* test: update expected params

* feat(nhost): add graphql-ws package

* fix(nhost): move index.ts to dataProvider folder

* fix(nhost): add liveProvider

* feat(nhost): add liveProvider to the example

* fix(graphql): export generete functions

* feat(graphql): add liveProvider support

* fix: update error messages

* chore: add changeset

* chore: update changeset links

* fix: Update changeset


* fix: update changeset


* docs: add graphql subscriptions promotions

* chore: add comment for graphql subscriptions support on hasura & nhost examples
