+++ .changeset/sour-owls-learn.md
@@ -1,0 +1,27 @@
+---
+"@pankod/refine-core": minor
+"@pankod/refine-graphql": minor
+"@pankod/refine-hasura": minor
+"@pankod/refine-nhost": minor
+---
+
+### `@pankod/refine-core`
+
+-   Added extra params to `useSubscription` and `useResourceSubscription`
+-   `useOne`, `useMany` and `useList` passed extra params to own subscription hook.
+
+### `@pankod/refine-hasura`
+
+-   Added `liveProvider`.
+
+To see an example of how to use it, check out [`here`](https://github.com/pankod/refine/blob/master/examples/dataProvider/hasura/src/App.tsx).
+
+### `@pankod/refine-nhost`
+
+-   Added `liveProvider`.
+
+To see an example of how to use it, check out [`here`](https://github.com/pankod/refine/blob/master/examples/dataProvider/nhost/src/App.tsx).
+
+### `@pankod/refine-graphql`
+
+-   Added `liveProvider`.
