+++ documentation/docs/core/providers/live-provider.md
@@ -27,6 +27,8 @@ const liveProvider = {
 -   **Ably** &#8594 [Source Code](https://github.com/pankod/refine/blob/master/packages/ably/src/index.ts) - [Demo](https://stackblitz.com/github/pankod/refine/tree/master/examples/ably/?preset=node)
 -   **Supabase** &#8594 [Source Code](https://github.com/pankod/refine/blob/master/packages/supabase/src/index.ts#L187)
 -   **Appwrite** &#8594 [Source Code](https://github.com/pankod/refine/blob/master/packages/appwrite/src/index.ts#L252)
+-   **Hasura** &#8594 [Source Code](https://github.com/pankod/refine/blob/feat/graphql-live-provider/packages/hasura/src/liveProvider/index.ts#L16)
+-   **Nhost** &#8594 [Source Code](https://github.com/pankod/refine/blob/feat/graphql-live-provider/packages/nhost/src/liveProvider/index.ts#L16)
 
 :::
