+++ packages/core/src/contexts/live/ILiveContext.ts
@@ -1,4 +1,11 @@
-import { BaseKey, LiveEvent } from "../../interfaces";
+import {
+    BaseKey,
+    CrudFilters,
+    CrudSorting,
+    LiveEvent,
+    MetaDataQuery,
+    Pagination,
+} from "../../interfaces";
 
 export type ILiveContext =
     | {
@@ -7,6 +14,14 @@ export type ILiveContext =
               channel: string;
               params?: {
                   ids?: BaseKey[];
+                  id?: BaseKey;
+                  metaData?: MetaDataQuery;
+                  pagination?: Pagination;
+                  hasPagination?: boolean;
+                  sort?: CrudSorting;
+                  filters?: CrudFilters;
+                  subscriptionType?: "useList" | "useOne" | "useMany";
+                  resource?: string;
                   [key: string]: any;
               };
               types: LiveEvent["type"][];
