+++ packages/core/src/hooks/data/useOne.ts
@@ -66,7 +66,13 @@ export const useOne = <
         resource,
         types: ["*"],
         channel: `resources/${resource}`,
-        params: { ids: id ? [id] : [], ...liveParams },
+        params: {
+            ids: id ? [id] : [],
+            id: id,
+            metaData,
+            subscriptionType: "useOne",
+            ...liveParams,
+        },
         enabled: queryOptions?.enabled,
         liveMode,
         onLiveEvent,
