+++ packages/hasura/package.json
@@ -34,12 +34,14 @@
     "@pankod/refine-core": "^3.38.2",
     "gql-query-builder": "^3.5.5",
     "graphql": "^15.6.1",
-    "graphql-request": "^4.3.0"
+    "graphql-request": "^4.3.0",
+    "graphql-ws": "^5.9.1"
   },
   "peerDependencies": {
     "@pankod/refine-core": "^3.23.2",
     "graphql-request": "^4.3.0",
-    "gql-query-builder": "^3.5.5"
+    "gql-query-builder": "^3.5.5",
+    "graphql-ws": "^5.9.1"
   },
   "repository": {
     "type": "git",
