+++ packages/nhost/package.json
@@ -34,11 +34,13 @@
     "@nhost/nhost-js": "^0.3.9",
     "@pankod/refine-core": "^3.36.0",
     "gql-query-builder": "^3.5.5",
-    "graphql": "^15.6.1"
+    "graphql": "^15.6.1",
+    "graphql-ws": "^5.9.1"
   },
   "peerDependencies": {
     "@pankod/refine-core": "^3.23.2",
-    "gql-query-builder": "^3.5.5"
+    "gql-query-builder": "^3.5.5",
+    "graphql-ws": "^5.9.1"
   },
   "repository": {
     "type": "git",
